Refactor file writing to use internal.AtomicWriteFile for improved data integrity and consistency; remove legacy utility functions and add new auth and output handling.
